New files for creating a project and a timestamp
