+++ index.html
@@ -11,5 +11,6 @@
     <div id="content"></div>
     <script src="requirements.js"></script>
     <script src="dist/js/app.js"></script>
+
   </body>
 </html>
