+++ src/js/index.js
@@ -2,11 +2,14 @@
   const ReactDOM = require('react-dom');
   const React = require('react');
   const TPane = require('./tpane');
-// var db = require('./db-init');
+  //var db = require('./db-init');
   const UploadModal = require('./uploadmodal');
   const remote = window.electron.remote;
   const {Menu} = remote;
   const menubar = require('./menubar');
+  const TimeStamp = require('./Timestamp');
+  const CheckLib = require('./Createproject');
+
 
   var App = {
     init: function() {
@@ -16,6 +19,8 @@
         <div>
           <TPane />
           <UploadModal />
+          <CheckLib />
+          <TimeStamp />
         </div>
       );
       ReactDOM.render(Application, document.getElementById('content'));
